refactor: cardQuantity -> CartQuantity
