--- src/component/@common/CardQuantity/index.tsx
@@ -1,7 +1,0 @@
-import { CartQuantityStyle } from './CardQuantity.styles';
-
-const CardQuantity = ({ children }: { children: React.ReactNode }) => {
-  return <div css={CartQuantityStyle}>{children}</div>;
-};
-
-export default CardQuantity;
